@@ -2,129 +2,88 @@
 
 namespace Bgaze\BootstrapForm;
 
-use Bgaze\BootstrapForm\Inputs;
-use Bgaze\BootstrapForm\Support\Attributes;
+use ArrayAccess;
+use Bgaze\BootstrapForm\Support\Html\Attributes;
+use Bgaze\BootstrapForm\Support\Html\Html;
 use Bgaze\BootstrapForm\Support\Traits\HasSettings;
-use Collective\Html\FormBuilder;
-use Collective\Html\HtmlBuilder;
+use Illuminate\Contracts\Session\Session;
 use Illuminate\Database\Eloquent\Model;
 use Illuminate\Support\Arr;
 use Illuminate\Support\Collection;
+use Illuminate\Support\Facades\Request;
+use Illuminate\Support\Facades\URL;
 use Illuminate\Support\Str;
 use Illuminate\Support\Traits\Macroable;
 
 /**
  * Form configuration:
  *
- * @property Model $model
+ * @property ?Model $model
  * @property string $error_bag
- * @property string $url
- * @property string $route
- * @property string $action
- * @property string $store
- * @property string $update
+ * @property ?bool $files
+ * @property ?string $url
+ * @property null|string|array $route
+ * @property null|string|array $action
+ * @property null|string|array $store
+ * @property null|string|array $update
  * @property string $layout
+ * @property bool $custom
+ * @property bool $show_all_errors
+ * @property ?array $group
  * @property string $left_class
  * @property string $right_class
- * @property bool $show_all_errors
- * @property array $group
+ * @property ?string $pull_right
+ * @property ?string $lspace
+ * @property ?string $hspace
+ * @property ?string $vspace
  */
 class BootstrapForm
 {
-
-    use Macroable;
     use HasSettings;
+    use Macroable;
 
     /**
      * Configuration keys that won't be inherited by form inputs.
      */
-    const RESERVED = ['model', 'url', 'route', 'action', 'update', 'store'];
-
-    /**
-     * Illuminate HtmlBuilder instance.
-     *
-     * @var HtmlBuilder
-     */
-    protected $html;
-
-    /**
-     * Illuminate FormBuilder instance.
-     *
-     * @var FormBuilder
-     */
-    protected $form;
+    const RESERVED = ['model', 'url', 'route', 'action', 'update', 'store', 'files'];
 
     /**
      * The form attribute set.
-     *
-     * @var Attributes
      */
-    protected $attributes;
+    protected Attributes $attributes;
 
     /**
-     * Whether a form is currently opened.
-     *
-     * @var Attributes
+     * Is a form opened ?
      */
-    protected $opened;
+    protected bool $form = false;
 
-
-    /**
-     * Constructor.
-     *
-     * @param  HtmlBuilder  $html
-     * @param  FormBuilder  $form
-     */
-    public function __construct(HtmlBuilder $html, FormBuilder $form)
+    public function __construct()
     {
-        $this->html = $html;
-        $this->form = $form;
         $this->resetForm();
     }
 
-    ### MISC UTILITIES #########################################################
-
+    // MISC UTILITIES
 
     /**
-     * Returns the "inheritable" forms settings.
-     *
-     * @return Collection
+     * Get inheritable forms settings.
      */
-    public function settings()
+    public function settings(): Collection
     {
         return $this->settings->except(static::RESERVED);
     }
 
-
     /**
-     * Returns the Illuminate HtmlBuilder instance.
-     *
-     * @return HtmlBuilder
+     * Get session store.
      */
-    public function htmlBuilder()
+    public function getSessionStore(): Session
     {
-        return $this->html;
+        return Request::session();
     }
 
 
-    /**
-     * Returns the Illuminate FormBuilder instance.
-     *
-     * @return FormBuilder
-     */
-    public function formBuilder()
-    {
-        return $this->form;
-    }
+    // FORM ELEMENT
 
-    ### CONFIGURATION ##########################################################
-
-
-    /**
-     * Reset form settings and attributes to default configuration.
-     *
-     */
-    protected function resetForm()
+    protected function resetForm(): void
     {
         // Set defaults options.
         $reserved = array_fill_keys(self::RESERVED, null);
@@ -139,20 +98,63 @@ class BootstrapForm
         }
 
         // Set default attributes.
-        $this->attributes = Attributes::make(['role' => 'form']);
+        $this->attributes = Attributes::make([
+            'role' => 'form',
+            'accept-charset' => 'utf-8',
+            'method' => 'POST'
+        ]);
 
         // Mark form as closed.
-        $this->opened = false;
+        $this->form = false;
     }
 
+    protected function getFormAction(): ?string
+    {
+        if ($this->url) {
+            return URL::to($this->url);
+        }
+
+        if ($this->model && !$this->route && !$this->action) {
+            $action = null;
+
+            if ($this->model->exists && $this->update) {
+                $action = [...Arr::wrap($this->update), $this->model->getRouteKey()];
+                $this->attributes->method = 'PUT';
+            } elseif (!$this->model->exists && $this->store) {
+                $action = Arr::wrap($this->store);
+                $this->attributes->method = 'POST';
+            }
+
+            if ($action) {
+                if (is_array(Arr::first($action)) || Str::contains(Arr::first($action), '@')) {
+                    $this->action = $action;
+                } else {
+                    $this->route = $action;
+                }
+            }
+        }
+
+        if ($action = $this->route ?: $this->action) {
+            $parameters = Arr::wrap($action);
+            $action = array_shift($parameters);
+
+            if ($this->route) {
+                return URL::route($action, $parameters);
+            }
+
+            return URL::action($action, $parameters);
+        }
+
+        return null;
+    }
 
     /**
-     * Init form with provided option set.
-     *
-     * @param  array  $options
+     * Open a form.
      */
-    protected function initForm(array $options = [])
+    public function open(array $options = []): string
     {
+        // Mark form as opened.
+        $this->form = true;
 
         // Merge with provided options.
         $settings = collect($options)->only($this->settings->keys())->except('group');
@@ -165,6 +167,11 @@ class BootstrapForm
             $this->group = array_merge($this->group, $options['group']);
         }
 
+        // Ignore model if not an instance of Illuminate\Database\Eloquent\Model
+        if (!$this->model instanceof Model) {
+            $this->model = null;
+        }
+
         // Set form attributes.
         $attributes = collect($options)->except($this->settings->keys())->except('group');
         $this->attributes = $this->attributes->merge($attributes);
@@ -174,522 +181,295 @@ class BootstrapForm
             $this->attributes->addClass('form-' . $this->layout);
         }
 
-        // Use explicity passed URL, route or action.
-        foreach (['url', 'route', 'action'] as $action) {
-            if ($this->{$action}) {
-                $this->attributes->{$action} = $this->{$action};
-                return;
-            }
-        }
-    }
-
-
-    /**
-     * Init Model related form settings.
-     *
-     * @param  Model  $model
-     */
-    protected function initModelForm(Model $model)
-    {
-        // Adjust form settings based on model existence.
-        if ($model->exists && $this->update) {
-            // Set form method.
-            $this->attributes->method = 'PUT';
-
-            // Set form action.
-            if (is_array($this->update)) {
-                $type = Str::contains(Arr::first($this->update), '@') ? 'action' : 'route';
-                $update = $this->update;
-                $update[] = $model->getRouteKey();
-                $this->attributes->{$type} = $update;
-            } else {
-                $type = Str::contains($this->update, '@') ? 'action' : 'route';
-                $this->attributes->{$type} = [$this->update, $model->getRouteKey()];
-            }
-        } elseif (!$model->exists && $this->store) {
-            // Set form method.
-            $this->attributes->method = 'POST';
-
-            // Set form action.
-            $name = is_array($this->store) ? Arr::first($this->store) : $this->store;
-            $type = Str::contains($name, '@') ? 'action' : 'route';
-            $this->attributes->{$type} = $this->store;
-        }
-    }
-
-    ### BUILD FORM ELEMENT #####################################################
-
-
-    /**
-     * Open a form.
-     *
-     * @param  array  $options
-     * @return string
-     */
-    public function open(array $options = [])
-    {
-        // Configure form.
-        $this->initForm($options);
-
-        // Mark form as opened.
-        $this->opened = true;
-
-        // If model, open model form.
-        if ($this->model instanceof Model) {
-            $this->initModelForm($this->model);
-            return $this->form->model($this->model, $this->attributes->toArray())->toHtml();
+        // Manage enctype
+        if ($this->files) {
+            $this->attributes->enctype = 'multipart/form-data';
         }
 
-        // Otherwise init standard form.
-        return $this->form->open($this->attributes->toArray())->toHtml();
-    }
+        // Set form action.
+        if ($action = $this->getFormAction()) {
+            $this->attributes->action = $action;
+        }
 
+        return Html::form($this->attributes->toArray())->open();
+    }
 
     /**
      * Reset and close the form.
-     *
-     * @return string
      */
-    public function close()
+    public function close(): string
     {
         $this->resetForm();
-        return $this->form->close()->toHtml();
-    }
 
+        return Html::form()->close();
+    }
 
     /**
      * Open a vertical Bootstrap form.
-     *
-     * @param  array  $options
-     * @return string
      */
-    public function vertical(array $options = [])
+    public function vertical(array $options = []): string
     {
-        $options['layout'] = 'vertical';
-        return $this->open($options);
+        return $this->open(['layout' => 'vertical'] + $options);
     }
-
 
     /**
      * Open an inline Bootstrap form.
-     *
-     * @param  array  $options
-     * @return string
      */
-    public function inline(array $options = [])
+    public function inline(array $options = []): string
     {
-        $options['layout'] = 'inline';
-        return $this->open($options);
+        return $this->open(['layout' => 'inline'] + $options);
     }
-
 
     /**
      * Open a horizontal Bootstrap form.
-     *
-     * @param  array  $options
-     * @return string
      */
-    public function horizontal(array $options = [])
+    public function horizontal(array $options = []): string
     {
-        $options['layout'] = 'horizontal';
-        return $this->open($options);
+        return $this->open(['layout' => 'horizontal'] + $options);
     }
 
-    ### BUILD FORM INPUTS ######################################################
-
+    // BUILD FORM INPUTS
 
     /**
      * Create a Bootstrap input by tag.
-     *
-     * @param  string  $tag
-     * @param  string  $name
-     * @param  string  $label
-     * @param  string  $value
-     * @param  array  $options
-     * @return string
      */
-    protected function input($tag, $name, $label = null, $value = null, array $options = [])
+    public function input(string $type, string $name, $label = null, $value = null, array $options = []): string
     {
-        $options['tag'] = $tag;
-        return Inputs\TextInput::make($name, $label, $value, $options);
+        return Inputs\TextInput::make($name, $label, $value, compact('type') + $options)->toHtml();
     }
 
-
     /**
-     * Create a Bootstrap text field input.
-     *
-     * @param  string  $name
-     * @param  string  $label
-     * @param  string  $value
-     * @param  array  $options
-     * @return string
+     * Create a Bootstrap text input.
      */
-    public function text($name, $label = null, $value = null, array $options = [])
+    public function text(string $name, $label = null, $value = null, array $options = []): string
     {
         return $this->input('text', $name, $label, $value, $options);
     }
 
-
     /**
-     * Create a Bootstrap email field input.
-     *
-     * @param  string  $name
-     * @param  string  $label
-     * @param  string  $value
-     * @param  array  $options
-     * @return string
+     * Create a Bootstrap email input.
      */
-    public function email($name = 'email', $label = null, $value = null, array $options = [])
+    public function email(string $name = 'email', $label = null, $value = null, array $options = []): string
     {
         return $this->input('email', $name, $label, $value, $options);
     }
 
-
     /**
-     * Create a Bootstrap URL field input.
-     *
-     * @param  string  $name
-     * @param  string  $label
-     * @param  string  $value
-     * @param  array  $options
-     * @return string
+     * Create a Bootstrap URL input.
      */
-    public function url($name, $label = null, $value = null, array $options = [])
+    public function url(string $name, $label = null, $value = null, array $options = []): string
     {
         return $this->input('url', $name, $label, $value, $options);
     }
 
-
     /**
-     * Create a Bootstrap tel field input.
-     *
-     * @param  string  $name
-     * @param  string  $label
-     * @param  string  $value
-     * @param  array  $options
-     * @return string
+     * Create a Bootstrap tel input.
      */
-    public function tel($name, $label = null, $value = null, array $options = [])
+    public function tel(string $name, $label = null, $value = null, array $options = []): string
     {
         return $this->input('tel', $name, $label, $value, $options);
     }
 
-
     /**
-     * Create a Bootstrap number field input.
-     *
-     * @param  string  $name
-     * @param  string  $label
-     * @param  string  $value
-     * @param  array  $options
-     * @return string
+     * Create a Bootstrap number input.
      */
-    public function number($name, $label = null, $value = null, array $options = [])
+    public function number(string $name, $label = null, $value = null, array $options = []): string
     {
         return $this->input('number', $name, $label, $value, $options);
     }
 
-
     /**
-     * Create a Bootstrap date field input.
-     *
-     * @param  string  $name
-     * @param  string  $label
-     * @param  string  $value
-     * @param  array  $options
-     * @return string
+     * Create a Bootstrap date input.
      */
-    public function date($name, $label = null, $value = null, array $options = [])
+    public function date(string $name, $label = null, $value = null, array $options = []): string
     {
         return $this->input('date', $name, $label, $value, $options);
     }
 
-
     /**
-     * Create a Bootstrap email time input.
-     *
-     * @param  string  $name
-     * @param  string  $label
-     * @param  string  $value
-     * @param  array  $options
-     * @return string
+     * Create a Bootstrap time input.
      */
-    public function time($name, $label = null, $value = null, array $options = [])
+    public function time(string $name, $label = null, $value = null, array $options = []): string
     {
         return $this->input('time', $name, $label, $value, $options);
     }
 
-
     /**
-     * Create a Bootstrap password field input.
-     *
-     * @param  string  $name
-     * @param  string  $label
-     * @param  array  $options
-     * @return string
+     * Create a Bootstrap password input.
      */
-    public function password($name, $label = null, array $options = [])
+    public function password(string $name, $label = null, array $options = []): string
     {
         return $this->input('password', $name, $label, null, $options);
     }
 
-
     /**
-     * Create a Bootstrap color field input.
-     *
-     * @param  string  $name
-     * @param  string  $label
-     * @param  array  $options
-     * @return string
+     * Create a Bootstrap color input.
      */
-    public function color($name, $label = null, $value = null, array $options = [])
+    public function color(string $name, $label = null, $value = null, array $options = []): string
     {
         return $this->input('color', $name, $label, $value, $options);
     }
 
-
     /**
-     * Create a Bootstrap textarea field input.
-     *
-     * @param  string  $name
-     * @param  string  $label
-     * @param  string  $value
-     * @param  array  $options
-     * @return string
+     * Create a Bootstrap textarea input.
      */
-    public function textarea($name, $label = null, $value = null, array $options = [])
+    public function textarea(string $name, $label = null, $value = null, array $options = []): string
     {
         return $this->input('textarea', $name, $label, $value, $options);
     }
 
-
     /**
      * Create a select box field.
-     *
-     * @param  string  $name
-     * @param  string  $label
-     * @param  array  $choices
-     * @param  string  $selected
-     * @param  array  $options
-     * @return string
      */
-    public function select($name, $label = null, $choices = [], $selected = null, array $options = [])
+    public function select(string $name, $label = null, ArrayAccess|array $choices = [], $selected = null, array $options = []): string
     {
-        return Inputs\SelectInput::make($name, $label, $choices, $selected, $options);
+        return Inputs\SelectInput::make($name, $label, $choices, $selected, $options)->toHtml();
     }
-
 
     /**
      * Create a Boostrap file upload button.
-     *
-     * @param  string  $name
-     * @param  string  $label
-     * @param  array  $options
-     * @return string
      */
-    public function file($name, $label = null, array $options = [])
+    public function file(string $name, $label = null, array $options = []): string
     {
-        return Inputs\FileInput::make($name, $label, $options);
+        return Inputs\FileInput::make($name, $label, $options)->toHtml();
     }
-
 
     /**
      * Create a Boostrap file upload button.
-     *
-     * @param  string  $name
-     * @param  string  $label
-     * @param  string  $value
-     * @param  array  $options
-     * @return string
      */
-    public function range($name, $label = null, $value = null, array $options = [])
+    public function range(string $name, $label = null, $value = null, array $options = []): string
     {
-        return Inputs\RangeInput::make($name, $label, $value, $options);
+        return Inputs\RangeInput::make($name, $label, $value, $options)->toHtml();
     }
 
+    /**
+     * Create a Bootstrap checkbox input.
+     */
+    public function checkbox(string $name, $label = null, $value = 1, $checked = null, array $options = []): string
+    {
+        return Inputs\CheckInput::make($name, $label, $value, $checked, ['type' => 'checkbox'] + $options)->toHtml();
+    }
+
+    /**
+     * Create a collection of Bootstrap checkboxes.
+     */
+    public function checkboxes(string $name, $label = null, array $choices = [], $checked = null, array $options = []): string
+    {
+        return Inputs\CheckChoice::make($name, $label, $choices, $checked, ['type' => 'checkbox'] + $options)->toHtml();
+    }
+
+    /**
+     * Create a Bootstrap radio input.
+     */
+    public function radio(string $name, $label = null, $value = null, $checked = null, array $options = []): string
+    {
+        return Inputs\CheckInput::make($name, $label, $value, $checked, ['type' => 'radio'] + $options)->toHtml();
+    }
+
+    /**
+     * Create a collection of Bootstrap radio inputs.
+     */
+    public function radios(string $name, $label = null, array $choices = [], $checked = null, array $options = []): string
+    {
+        $options['tag'] = 'radio';
+
+        return Inputs\CheckChoice::make($name, $label, $choices, $checked, ['type' => 'radio'] + $options)->toHtml();
+    }
 
     /**
      * Create a hidden field.
-     *
-     * @param  string  $name
-     * @param  string  $value
-     * @param  array  $options
-     * @return string
      */
-    public function hidden($name, $value = null, $options = [])
+    public function hidden(string $name, $value = null, array $options = []): string
     {
         if (!isset($options['id'])) {
             $options['id'] = $this->flattenName($name, '_');
         }
 
-        return $this->form->hidden($name, $value, $options);
+
+        return Html::input($options)
+            ->attributes([
+                'type' => 'hidden',
+                'name' => $name,
+                'value' => $value
+            ])
+            ->toHtml();
     }
 
-
-    /**
-     * Create a Bootstrap checkbox input.
-     *
-     * @param  string  $name
-     * @param  string  $label
-     * @param  mixed  $value
-     * @param  bool  $checked
-     * @param  array  $options
-     * @return string
-     */
-    public function checkbox($name, $label = null, $value = 1, $checked = null, array $options = [])
-    {
-        $options['tag'] = 'checkbox';
-        return Inputs\CheckInput::make($name, $label, $value, $checked, $options);
-    }
-
-
-    /**
-     * Create a collection of Bootstrap checkboxes.
-     *
-     * @param  string  $name
-     * @param  string  $label
-     * @param  array  $choices
-     * @param  mixed  $checked
-     * @param  array  $options
-     * @return string
-     */
-    public function checkboxes($name, $label = null, array $choices = [], $checked = null, array $options = [])
-    {
-        $options['tag'] = 'checkbox';
-        return Inputs\CheckChoice::make($name, $label, $choices, $checked, $options);
-    }
-
-
-    /**
-     * Create a Bootstrap radio input.
-     *
-     * @param  string  $name
-     * @param  string  $label
-     * @param  mixed  $value
-     * @param  bool  $checked
-     * @param  array  $options
-     * @return string
-     */
-    public function radio($name, $label = null, $value = null, $checked = null, array $options = [])
-    {
-        $options['tag'] = 'radio';
-        return Inputs\CheckInput::make($name, $label, $value, $checked, $options);
-    }
-
-
-    /**
-     * Create a collection of Bootstrap radio inputs.
-     *
-     * @param  string  $name
-     * @param  string  $label
-     * @param  array  $choices
-     * @param  mixed  $checked
-     * @param  array  $options
-     * @return string
-     */
-    public function radios($name, $label = null, array $choices = [], $checked = null, array $options = [])
-    {
-        $options['tag'] = 'radio';
-        return Inputs\CheckChoice::make($name, $label, $choices, $checked, $options);
-    }
-
-    ### BUILD MISC ELEMENTS ####################################################
-
+    // MISC
 
     /**
      * Create a Bootstrap label.
-     *
-     * @param  string  $name
-     * @param  string  $value
-     * @param  array  $options
-     * @return string
      */
-    public function label($name, $value = null, array $options = [], $escapeHtml = false)
+    public function label(string $for, $content = null, array $options = []): string
     {
-        return $this->form->label($name, $value, $options, $escapeHtml);
+        return Html::label($options)->attribute('for', $for)->append($content)->toHtml();
     }
 
-
     /**
-     * Create options for Boostrap styled button.
-     *
-     * @param  string  $style
-     * @param  mixed  $options
-     * @return array
+     * Create options for Boostrap button.
      */
-    protected function buttonOption($style, $options)
+    protected function btn(string $style, array|string|null $options): array
     {
-        $class = '';
+        $attributes = Attributes::make()
+            ->addClass('btn')
+            ->addClass(sprintf('btn-%s', (!empty($options) && is_string($options)) ? $options : $style));
 
-        if ($this->opened && $this->layout === 'inline') {
+        if ($this->form && $this->layout === 'inline') {
             if ($this->hspace) {
-                $class .= $this->hspace . ' ';
+                $attributes->addClass($this->hspace);
             }
 
             if ($this->vspace) {
-                $class .= $this->vspace . ' ';
+                $attributes->addClass($this->vspace);
             }
         }
 
-        $class .= 'btn btn-';
-
-        if (is_array($options)) {
-            return array_merge(['class' => $class . $style], $options);
+        if (!empty($options) && !is_string($options)) {
+            $attributes = $attributes->merge($options);
         }
 
-        return ['class' => $class . (empty($options) ? $style : $options)];
+        return $attributes->toArray();
     }
-
 
     /**
-     * Create a Boostrap submit button.
-     *
-     * @param  string  $value
-     * @param  mixed  $options
-     * @return string
+     * Create a Boostrap submit input.
      */
-    public function submit($value = null, $options = null)
+    public function submit($value = null, array|string|null $options = null): string
     {
-        return $this->form->submit($value, $this->buttonOption('primary', $options));
+        return Html::input($this->btn('primary', $options))
+            ->attributes(['type' => 'submit', 'value' => $value])
+            ->toHtml();
     }
-
 
     /**
-     * Create a Boostrap reset button.
-     *
-     * @param  string  $value
-     * @param  mixed  $options
-     * @return string
+     * Create a Boostrap reset input.
      */
-    public function reset($value = null, $options = null)
+    public function reset($value = null, array|string|null $options = null): string
     {
-        return $this->form->reset($value, $this->buttonOption('danger', $options));
+        return Html::input($this->btn('danger', $options))
+            ->attributes(['type' => 'reset', 'value' => $value])
+            ->toHtml();
     }
-
 
     /**
      * Create a Boostrap button.
-     *
-     * @param  string  $value
-     * @param  mixed  $options
-     * @return string
      */
-    public function button($value = null, $options = null)
+    public function button($content = null, array|string|null $options = null): string
     {
-        return $this->form->button($value, $this->buttonOption('primary', $options));
+        return Html::button(['type' => 'button'])
+            ->attributes($this->btn('primary', $options))
+            ->append($content)
+            ->toHtml();
     }
-
 
     /**
      * Create a Boostrap link button.
-     *
-     * @param  string  $url
-     * @param  string  $title
-     * @param  mixed  $options
-     * @return string
      */
-    public function link($url, $title = null, $options = null)
+    public function link(string $href, $content = null, array|string|null $options = null): string
     {
-        return $this->html->link($url, $title, $this->buttonOption('primary', $options), null, false);
+        return Html::a($this->btn('primary', $options))
+            ->attribute('href', $href)
+            ->append($content)
+            ->toHtml();
     }
 }
