@@ -7,72 +7,41 @@ use Illuminate\Support\ServiceProvider;
 
 class BootstrapFormServiceProvider extends ServiceProvider
 {
+    protected bool $defer = true;
 
-    /**
-     * Indicates if loading of the provider is deferred.
-     *
-     * @var bool
-     */
-    protected $defer = true;
-
-
-    /**
-     * Register the service provider.
-     */
-    public function register()
+    public function register(): void
     {
-        // Merge configuration.
         $this->mergeConfigFrom(__DIR__ . '/config/config.php', 'bootstrap_form');
 
-        // Register service.
-        $this->app->singleton('bootstrap_form', function ($app) {
-            return new BootstrapForm($app['html'], $app['form']);
-        });
+        $this->app->singleton('bootstrap_form', fn($app) => new BootstrapForm());
     }
 
 
-    /**
-     * Boot the service provider.
-     */
-    public function boot()
+    public function boot(): void
     {
-        // Publish configuration.
         $this->publishes([__DIR__ . '/config/config.php' => config_path('bootstrap_form.php')], 'config');
 
-        // Register blade directives if enabled.
         if (config('bootstrap_form.blade_directives', true)) {
             $this->registerBladeDirectives();
         }
     }
 
-
-    /**
-     * Register blade directive.
-     */
-    protected function registerBladeDirectives()
+    protected function registerBladeDirectives(): void
     {
         $functions = [
             'open', 'close', 'vertical', 'inline', 'horizontal',
-            'text', 'email', 'url', 'tel', 'number', 'date', 'time', 'textarea', 'password','color',
+            'text', 'email', 'url', 'tel', 'number', 'date', 'time', 'textarea', 'password', 'color',
             'file', 'hidden', 'select', 'range',
             'checkbox', 'checkboxes', 'radio', 'radios',
             'label', 'submit', 'reset', 'button', 'link',
         ];
 
         foreach ($functions as $f) {
-            Blade::directive($f, function ($expression) use ($f) {
-                return "<?= BF::{$f}({$expression}); ?>";
-            });
+            Blade::directive($f, fn($expression) => sprintf('<?= BF::%s(%s) ?>', $f, $expression));
         }
     }
 
-
-    /**
-     * Get the services provided by the provider.
-     *
-     * @return array
-     */
-    public function provides()
+    public function provides(): array
     {
         return ['bootstrap_form'];
     }
