@@ -3,155 +3,101 @@
 namespace Bgaze\BootstrapForm\Support;
 
 use Bgaze\BootstrapForm\Support\Facades\BF;
+use Bgaze\BootstrapForm\Support\Html\Attributes;
+use Bgaze\BootstrapForm\Support\Html\Html;
 use Bgaze\BootstrapForm\Support\Traits\HasSettings;
-use Collective\Html\FormBuilder;
-use Collective\Html\HtmlBuilder;
+use Illuminate\Contracts\Support\Htmlable;
+use Illuminate\Contracts\Support\Renderable;
 use Illuminate\Support\Collection;
 use Illuminate\Support\Str;
+use Stringable;
 
 /**
  * Input settings:
  *
- * @property string $errors
+ * @property ?string $errors
  * @property string $name
- * @property string $value
- * @property mixed $label
- * @property string $help
+ * @property $value
+ * @property $label
+ * @property $help
  *
  * Inherited from form:
- *
+ * @property ?array $group
  * @property string $layout
  * @property string $error_bag
  * @property bool $show_all_errors
- * @property bool $pull_right
+ * @property ?string $pull_right
  * @property string $left_class
  * @property string $right_class
- * @property string $spacer
- * @property bool|array $group
+ * @property ?string $lspace
+ * @property ?string $hspace
+ * @property ?string $vspace
  */
-abstract class Input
+abstract class Input implements Htmlable, Stringable, Renderable
 {
-
     use HasSettings;
 
-    /**
-     * Illuminate HtmlBuilder instance.
-     *
-     * @var HtmlBuilder
-     */
-    protected $html;
+    protected Attributes $input_attributes;
 
-    /**
-     * Illuminate FormBuilder instance.
-     *
-     * @var FormBuilder
-     */
-    protected $form;
+    protected Attributes $label_attributes;
 
-    /**
-     * The input attributes repository.
-     *
-     * @var Attributes
-     */
-    protected $input_attributes;
+    protected Attributes $group_attributes;
 
-    /**
-     * The label attributes repository.
-     *
-     * @var Attributes
-     */
-    protected $label_attributes;
-
-    /**
-     * The group attributes repository.
-     *
-     * @var Attributes
-     */
-    protected $group_attributes;
-
-
-    /**
-     * The class constructor.
-     *
-     * @param  string  $name
-     * @param  mixed  $label
-     * @param  mixed  $value
-     * @param  array  $options
-     */
-    public function __construct($name, $label = null, $value = null, array $options = [])
+    public function __construct(string $name, $label = null, $value = null, array $options = [])
     {
-        // Resolve html builders.
-        $this->html = BF::htmlBuilder();
-        $this->form = BF::formBuilder();
-
-        // Get field configuration.
         $this->configure($name, $label, $value, $options);
         $this->setInputAttributes($options);
         $this->setLabelAttributes($options);
         $this->setGroupAttributes($options);
-        $this->getErrors();
+        $this->setErrors();
     }
 
-
-    /**
-     * Return the input as a HTML string.
-     *
-     * @return string
-     */
-    public function __toString()
+    public function toHtml(): string
     {
         return $this->group();
     }
 
-    ### CONFIGURATION ##########################################################
+    public function render(): string
+    {
+        return $this->toHtml();
+    }
 
+    public function __toString(): string
+    {
+        return $this->toHtml();
+    }
 
-    /**
-     * Get the input default options.
-     *
-     * @return Collection
-     */
-    protected function defaults()
+    // CONFIGURATION
+
+    protected function defaults(): Collection
     {
         return BF::settings()->merge([
             'help' => false,
         ]);
     }
 
-
-    /**
-     * Merge provided options with default configuration.
-     *
-     * @param  string  $name
-     * @param  mixed  $label
-     * @param  mixed  $value
-     * @param  array  $options
-     */
-    protected function configure($name, $label, $value, array $options)
+    protected function configure(string $name, $label = null, $value = null, array $options = []): void
     {
         // Get default settings.
         $this->settings = static::defaults();
 
         // Merge with provided configuration.
-        $settings = collect($options)
-            ->only($this->settings->keys())
-            ->except('group');
+        $settings = collect($options)->only($this->settings->keys())->except('group');
         $this->settings = $this->settings->merge($settings);
 
         // Add reserved values.
         $this->name = $name;
-        $this->value = $value;
+        $this->value = $this->setValue($value);
         $this->label = $label;
         $this->errors = '';
     }
 
+    protected function setValue($value)
+    {
+        return $value;
+    }
 
-    /**
-     * Set input attributes.
-     *
-     * @param  array  $options
-     */
-    protected function setInputAttributes(array $options)
+    protected function setInputAttributes(array $options): void
     {
         $this->input_attributes = Attributes::make($options)->except($this->settings->keys());
 
@@ -160,13 +106,7 @@ abstract class Input
         }
     }
 
-
-    /**
-     * Set label attributes.
-     *
-     * @param  array  $options
-     */
-    protected function setLabelAttributes(array $options)
+    protected function setLabelAttributes(array $options): void
     {
         $label = $this->label;
 
@@ -179,15 +119,15 @@ abstract class Input
         } else {
             $this->label_attributes = Attributes::make();
         }
+
+        $this->label_attributes->for = $this->input_attributes->id;
+
+        if ($this->layout === 'inline' && $this->lspace) {
+            $this->label_attributes->addClass($this->lspace);
+        }
     }
 
-
-    /**
-     * Set group attributes.
-     *
-     * @param  array  $options
-     */
-    protected function setGroupAttributes(array $options)
+    protected function setGroupAttributes(array $options): void
     {
         if (isset($options['group'])) {
             if ($options['group'] === false) {
@@ -209,111 +149,6 @@ abstract class Input
         if (!$this->group_attributes->id) {
             $this->group_attributes->id = $this->flattenName($this->name, '_') . '_group';
         }
-    }
-
-
-    /**
-     * Get the template to build up error messages.
-     */
-    protected function errorTemplate()
-    {
-        return '<div class="invalid-feedback">:message</div>';
-    }
-
-
-    /**
-     * Compile inputs errors to and HTML string.
-     * Add error class to input and group if there is any errors.
-     */
-    protected function getErrors()
-    {
-        $errors = $this->form->getSessionStore()->get('errors');
-        if (empty($errors)) {
-            return;
-        }
-
-        $errorBag = $errors->{$this->error_bag} ?? false;
-        if (!$errorBag) {
-            return;
-        }
-
-        $field = $this->flattenName($this->name, '.');
-        if (!$errorBag->has($field)) {
-            return;
-        }
-
-        if ($this->show_all_errors) {
-            $this->errors = implode('', $errorBag->get($field, $this->errorTemplate()));
-        } else {
-            $this->errors = $errorBag->first($field, $this->errorTemplate());
-        }
-
-        $this->input_attributes->addClass('is-invalid');
-        $this->group_attributes->addClass('is-invalid');
-    }
-
-    ### COMPONENTS #############################################################
-
-
-    /**
-     * Compile input to a HTML string.
-     *
-     * @return string
-     */
-    public abstract function input();
-
-
-    /**
-     * Decorate the input to get the final Bootstrap format.
-     *
-     * @return string
-     */
-    public function inputGroup()
-    {
-        return $this->input();
-    }
-
-
-    /**
-     * Compile label to a HTML string.
-     *
-     * @return string
-     */
-    public function label()
-    {
-        if ($this->label === false) {
-            return '';
-        }
-
-        return $this->form->label($this->input_attributes->id, $this->label, $this->label_attributes->toArray(), false)->toHtml();
-    }
-
-
-    /**
-     * Compile label to a HTML string.
-     *
-     * @return string
-     */
-    public function help()
-    {
-        if ($this->help === false) {
-            return '';
-        }
-
-        return $this->html->tag('small', $this->help, ['class' => 'form-text'])->toHtml();
-    }
-
-
-    /**
-     * Compile group to a HTML string.
-     *
-     * @return string
-     */
-    public function group()
-    {
-        if (!$this->group) {
-            return $this->inputGroup();
-        }
 
         $this->group_attributes->addClass('form-group');
 
@@ -328,22 +163,66 @@ abstract class Input
         if ($this->layout === 'inline' && $this->vspace) {
             $this->group_attributes->addClass($this->vspace);
         }
-
-        $content = $this->leftGroupColumn() . $this->rightGroupColumn();
-
-        return $this->html->tag('div', $content, $this->group_attributes->toArray())->toHtml();
     }
 
+    protected function errorTemplate(): string
+    {
+        return '<div class="invalid-feedback d-block">:message</div>';
+    }
 
-    /**
-     * Get the "left" part of the form group (label or spacer in "pull right" mode).
-     *
-     * @return string
-     */
-    protected function leftGroupColumn()
+    protected function setErrors(): void
+    {
+        $this->errors = null;
+
+        if ($errorBag = BF::getSessionStore()->get('errors')?->{$this->error_bag}) {
+            $field = $this->flattenName($this->name, '.');
+            if ($errorBag->has($field)) {
+                if ($this->show_all_errors) {
+                    $this->errors = implode('', $errorBag->get($field, $this->errorTemplate()));
+                } else {
+                    $this->errors = $errorBag->first($field, $this->errorTemplate());
+                }
+
+                $this->input_attributes->addClass('is-invalid');
+                $this->group_attributes->addClass('is-invalid');
+            }
+        }
+    }
+
+    // COMPONENTS
+
+    abstract public function input(): string;
+
+    public function inputGroup(): string
+    {
+        return $this->input();
+    }
+
+    public function label(): ?string
+    {
+        return ($this->label === false) ? null : Html::label($this->label_attributes->toArray())->append($this->label)->toHtml();
+    }
+
+    public function help(): ?string
+    {
+        return ($this->help === false) ? null : Html::small()->addClass('form-text')->append($this->help)->toHtml();
+    }
+
+    public function group(): string
+    {
+        if ($this->group) {
+            return Html::div($this->group_attributes->toArray())
+                ->append([$this->leftGroupColumn(), $this->rightGroupColumn()])
+                ->toHtml();
+        }
+
+        return $this->inputGroup();
+    }
+
+    protected function leftGroupColumn(): ?string
     {
         if ($this->layout === 'horizontal' && $this->pull_right && !$this->label) {
-            return $this->html->tag('div', '', ['class' => $this->left_class])->toHtml();
+            return Html::div()->addClass($this->left_class);
         }
 
         if ($this->layout === 'horizontal') {
@@ -357,24 +236,14 @@ abstract class Input
         return $this->label();
     }
 
-
-    /**
-     * Get the "right" part of the form group (input + errors + help wrapped in a div).
-     *
-     * @return string
-     */
-    protected function rightGroupColumn()
+    protected function rightGroupColumn(): string
     {
-        $content = $this->inputGroup();
-        $content .= $this->errors;
-        $content .= $this->help();
-
-        $attributes = Attributes::make();
+        $div = Html::div()->append([$this->inputGroup(), $this->errors, $this->help()]);
 
         if ($this->layout === 'horizontal') {
-            $attributes->addClass($this->pull_right ? 'col' : $this->right_class);
+            $div->addClass($this->pull_right ? 'col' : $this->right_class);
         }
 
-        return $this->html->tag('div', $content, $attributes->toArray())->toHtml();
+        return $div;
     }
 }
