@@ -3,26 +3,21 @@
 namespace Bgaze\BootstrapForm\Inputs;
 
 use Bgaze\BootstrapForm\Support\Input;
-use Bgaze\BootstrapForm\Inputs\CheckInput;
+use Illuminate\Support\Arr;
+use Illuminate\Support\Collection;
 
 /**
  * Specific settings:
- * 
+ *
  * @property string $tag
- * @property array  $choices
- * @property bool   $inline
- * @property bool   $custom
- * @property bool   $switch
+ * @property array $choices
+ * @property bool $inline
+ * @property bool $custom
+ * @property bool $switch
  */
 class CheckChoice extends Input
 {
-
-    /**
-     * Get the input default options.
-     *
-     * @return Collection
-     */
-    protected function defaults()
+    protected function defaults(): Collection
     {
         return parent::defaults()->merge([
             'tag' => 'checkbox',
@@ -32,42 +27,23 @@ class CheckChoice extends Input
         ]);
     }
 
-    /**
-     * The class constructor.
-     * 
-     * @param string $name
-     * @param mixed  $label
-     * @param array  $choices
-     * @param mixed  $checked
-     * @param array  $options
-     */
-    public function __construct($name, $label = null, array $choices = [], $checked = [], array $options = [])
+    public function __construct(string $name, $label = null, array $choices = [], $checked = [], array $options = [])
     {
         parent::__construct($name, $label, $checked, $options);
         $this->choices = $choices;
     }
 
-    /**
-     * Instanciate a CheckChoice.
-     * 
-     * @param string $name
-     * @param mixed  $label
-     * @param array  $choices
-     * @param mixed  $checked
-     * @param array  $options
-     * @return CheckChoice
-     */
-    public static function make($name, $label = null, array $choices = [], $checked = [], array $options = [])
+    public static function make(string $name, $label = null, array $choices = [], $checked = [], array $options = []): static
     {
         return new static($name, $label, $choices, $checked, $options);
     }
 
-    /**
-     * Set label attributes.
-     * 
-     * @param array $options 
-     */
-    protected function setLabelAttributes(array $options)
+    protected function setValue($value): array
+    {
+        return Arr::wrap(parent::setValue($value));
+    }
+
+    protected function setLabelAttributes(array $options): void
     {
         parent::setLabelAttributes($options);
 
@@ -76,45 +52,33 @@ class CheckChoice extends Input
         }
     }
 
-    /**
-     * Get the template to build up error messages.
-     */
-    protected function errorTemplate()
+    public function input(): string
     {
-        return '<div class="invalid-feedback d-block">:message</div>';
+        return collect($this->choices)
+            ->map(fn($label, $value) => (string)$this->makeCheckInput($value, $label))
+            ->join('');
     }
 
-    /**
-     * Compile input to a HTML string.
-     *
-     * @return string
-     */
-    public function input()
+    protected function makeCheckInput($value, $label): CheckInput
     {
-        $inputs = '';
+        $checked = in_array($value, $this->value);
 
-        foreach ($this->choices as $value => $label) {
-            $checked = in_array($value, (array) $this->value);
+        $options = $this->settings
+            ->except(['choices', 'name', 'value', 'label'])
+            ->merge($this->input_attributes)
+            ->merge([
+                'layout' => 'vertical',
+                'disable_errors' => true,
+                'group' => false,
+                'help' => false,
+                'id' => $this->flattenName($this->name, '_') . '_' . $value,
+            ])
+            ->toArray();
 
-            $options = $this->settings
-                ->except(['choices', 'name', 'value', 'label'])
-                ->merge($this->input_attributes)
-                ->merge([
-                    'layout' => 'vertical',
-                    'disable_errors' => true,
-                    'group' => false,
-                    'help' => false,
-                    'id' => $this->flattenName($this->name, '_') . '_' . $value,
-                ])
-                ->toArray();
-
-            if ($this->layout === 'inline') {
-                $options['inline'] = true;
-            }
-
-            $inputs .= CheckInput::make($this->name, $label, $value, $checked, $options);
+        if ($this->layout === 'inline') {
+            $options['inline'] = true;
         }
 
-        return $inputs;
+        return CheckInput::make($this->name, $label, $value, $checked, $options);
     }
 }
