@@ -2,30 +2,26 @@
 
 namespace Bgaze\BootstrapForm\Inputs;
 
+use Bgaze\BootstrapForm\Support\Html\Html;
 use Bgaze\BootstrapForm\Support\Input;
+use Illuminate\Support\Collection;
 
 /**
  * Specific settings:
- * 
- * @property string $tag
- * @property bool   $disable_errors
- * @property bool   $checked
- * @property bool   $inline
- * @property bool   $custom
- * @property bool   $switch
+ *
+ * @property string $type
+ * @property bool $disable_errors
+ * @property bool $checked
+ * @property bool $inline
+ * @property bool $custom
+ * @property bool $switch
  */
 class CheckInput extends Input
 {
-
-    /**
-     * Get the input default options.
-     *
-     * @return Collection
-     */
-    protected function defaults()
+    protected function defaults(): Collection
     {
         return parent::defaults()->merge([
-            'tag' => 'checkbox',
+            'type' => 'checkbox',
             'disable_errors' => false,
             'checked' => null,
             'inline' => false,
@@ -33,49 +29,25 @@ class CheckInput extends Input
         ]);
     }
 
-    /**
-     * The class constructor.
-     * 
-     * @param string $name
-     * @param mixed  $label
-     * @param mixed  $value
-     * @param mixed  $checked
-     * @param array  $options
-     */
-    public function __construct($name, $label = null, $value = 1, $checked = null, array $options = [])
+    public function __construct(string $name, $label = null, $value = 1, $checked = null, array $options = [])
     {
         parent::__construct($name, $label, $value, $options);
         $this->checked = $checked;
     }
 
-    /**
-     * Instanciate a CheckInput.
-     * 
-     * @param string $name
-     * @param mixed  $label
-     * @param mixed  $value
-     * @param mixed  $checked
-     * @param array  $options
-     * @return CheckInput
-     */
-    public static function make($name, $label = null, $value = 1, $checked = null, array $options = [])
+    public static function make(string $name, $label = null, $value = 1, $checked = null, array $options = []): static
     {
         return new static($name, $label, $value, $checked, $options);
     }
 
-    /**
-     * Set input attributes.
-     * 
-     * @param array  $options 
-     */
-    protected function setInputAttributes(array $options)
+    protected function setInputAttributes(array $options): void
     {
         parent::setInputAttributes($options);
 
-        if ($this->tag === 'radio') {
+        if ($this->type === 'radio') {
             $this->switch = false;
         } else {
-            $this->tag = 'checkbox';
+            $this->type = 'checkbox';
         }
 
         if ($this->switch) {
@@ -85,12 +57,7 @@ class CheckInput extends Input
         $this->input_attributes->addClass($this->custom ? 'custom-control-input' : 'form-check-input');
     }
 
-    /**
-     * Set label attributes.
-     * 
-     * @param array $options 
-     */
-    protected function setLabelAttributes(array $options)
+    protected function setLabelAttributes(array $options): void
     {
         parent::setLabelAttributes($options);
 
@@ -101,78 +68,57 @@ class CheckInput extends Input
         }
     }
 
-    /**
-     * Compile input to a HTML string.
-     *
-     * @return string
-     */
-    public function input()
+    public function input(): string
     {
-        return $this->form->{$this->tag}($this->name, $this->value, $this->checked, $this->input_attributes->toArray())->toHtml();
+        return Html::input($this->input_attributes->toArray())
+            ->attribute('type', $this->type)
+            ->attribute('value', $this->value)
+            ->attribute('checked', $this->checked)
+            ->toHtml();
     }
 
-    /**
-     * Decorate the input to get the final Bootstrap format.
-     *
-     * @return string
-     */
-    public function inputGroup()
+    public function inputGroup(): string
     {
-        $content = $this->input();
-        $content .= $this->label();
+        $div = Html::div()
+            ->append([$this->input(), $this->label()])
+            ->addClass($this->custom ? 'custom-control' : 'form-check');
 
-        if (!$this->disable_errors) {
-            $content .= $this->errors;
+        if ($this->errors && !$this->disable_errors) {
+            $div->append($this->errors);
         }
 
-        if ($this->help) {
-            $content .= $this->html->tag('small', $this->help, ['class' => 'form-text']);
+        if ($help = $this->help()) {
+            $div->append($help);
         }
-
-        $class = $this->custom ? 'custom-control' : 'form-check';
 
         if ($this->inline) {
-            $class .= $this->custom ? ' custom-control-inline' : ' form-check-inline';
+            $div->addClass($this->custom ? ' custom-control-inline' : ' form-check-inline');
         }
 
         if ($this->switch) {
-            $class .= ' custom-switch';
+            $div->addClass(' custom-switch');
         } elseif ($this->custom) {
-            $class .= " custom-{$this->tag}";
+            $div->addClass(" custom-{$this->type}");
         }
 
-        return $this->html->tag('div', $content, ['class' => $class])->toHtml();
+        return $div->toHtml();
     }
 
-    /**
-     * Compile label to a HTML string.
-     * 
-     * @return string
-     */
-    public function label()
+    public function label(): ?string
     {
         if ($this->label === false && $this->custom) {
-            return sprintf($this->form->label($this->input_attributes->id, '%s', $this->label_attributes->toArray(), false), '');
+            return Html::label($this->label_attributes->toArray())->attribute('for', $this->input_attributes->id);
         }
 
-        if ($this->label === false) {
-            return '';
-        }
-
-        return $this->form->label($this->input_attributes->id, $this->label, $this->label_attributes->toArray(), false)->toHtml();
+        return parent::label();
     }
 
-    /**
-     * Get the "left" part of the form group (empty or spacer in "pull right" mode).
-     * 
-     * @return string
-     */
-    protected function leftGroupColumn()
+    protected function leftGroupColumn(): ?string
     {
         if ($this->layout === 'horizontal' && $this->pull_right) {
-            return $this->html->tag('div', '', ['class' => $this->left_class])->toHtml();
+            return Html::div()->addClass($this->left_class);
         }
 
-        return '';
+        return null;
     }
 }
